pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Agentic RAG system with ADK, OpenAI, Milvus, Postgres, and BGE-M3"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "python-dotenv>=1.0",
  "pydantic>=2.6",
  "httpx[socks]>=0.27",
  "openai>=1.40",
  "google-adk>=0.1.0",
  "pymilvus>=2.4.0",
  "psycopg[binary]>=3.1",
  "SQLAlchemy>=2.0",
  "alembic>=1.13",
  "pypdf>=4.2",
  "python-docx>=1.1",
  "tiktoken>=0.7",
  "sentence-transformers>=2.7",
  "torch>=2.6",
  "FlagEmbedding>=1.2",
  "numpy>=1.26",
  "transformers==4.44.2",
  "ag-ui-protocol>=0.1.0",
]

[tool.uv]
managed = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "embedding_service"]
